requirements.txt
streamlit 
google-generativeai 
langchain 
langchain-community 
python-dotenv
langchain_google_genai